Create spiral.py
In this code, we created a square spiral pattern using the turtle "Squary." It uses the default colour, which is black. Using the command *squary.color("red")*, you can also alter the colour. You can substitute the desired colour for "red". The turtle forms squares into a spiral by going ahead and rotating left for 91 degrees, then doing this 500 times.