Create InsertionSort.py
Insertion Sort Implementation In Python
In order to rank an array of size N ascendingly:
Over the array, iterate from arr[1] to arr[N].
Comparing the current (key) element to the previous one.
Compare the key element to the previous elements to see if it is smaller. In order to make room for the substituted element, move the larger elements up one position.