Took out double turtle and extra line
I hid the extra turtle that was stationary in the middle of the flag until covered up by the circle that is created later on. Also, the script creates an orange line on the left side of the flag. I lifted the pen during this time to remove that line.